# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Rental Api",
        "description": "Demo Car Rental Api",
        "version": "1.0"
    },
    "paths": {
        "/car-rental/api/v1/bookings": {
            "post": {
                "tags": [
                    "bookings"
                ],
                "summary": "Create a car booking",
                "operationId": "158316d3a1524549ebc160af1511203c",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "properties": {
                                "token": {
                                    "type": "string",
                                    "example": "10293182301230123"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "car_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "example": "2023-04-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "example": "2023-04-05"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "booking_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "car_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "start_date": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-04-01T00:00:00.000000Z"
                                        },
                                        "end_date": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-04-05T00:00:00.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-03-18T09:25:53.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-03-18T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized: Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Invalid credentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/car-rental/api/v1/bookings/{id}": {
            "get": {
                "tags": [
                    "bookings"
                ],
                "summary": "Get a specific car booking",
                "operationId": "d0cf74e3267add49121c4ea03ea760b3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "properties": {
                                "token": {
                                    "type": "string",
                                    "example": "10293182301230123"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "booking_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "car_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "start_date": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-04-01T00:00:00.000000Z"
                                        },
                                        "end_date": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-04-05T00:00:00.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-03-18T09:25:53.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-03-18T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized: Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Invalid credentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "bookings"
                ],
                "summary": "Update a car booking",
                "operationId": "10b366f6bcbf3ec88aca732eb960f461",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "properties": {
                                "token": {
                                    "type": "string",
                                    "example": "10293182301230123"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "start_date": {
                                        "type": "string",
                                        "example": "2023-04-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "example": "2023-04-05"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "booking_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "car_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "start_date": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-04-01T00:00:00.000000Z"
                                        },
                                        "end_date": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-04-05T00:00:00.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-03-18T09:25:53.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2023-03-18T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized: Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Invalid credentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "bookings"
                ],
                "summary": "Delete a car booking",
                "operationId": "a803bf5441fd1ec2bda8acee7972a299",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "properties": {
                                "token": {
                                    "type": "string",
                                    "example": "10293182301230123"
                                }
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "deleted"
                    },
                    "403": {
                        "description": "Unauthorized: Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Invalid credentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}